% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qbit_metadata.R
\name{create_qbit_metadata}
\alias{create_qbit_metadata}
\title{Create Metatdata for QBits}
\usage{
create_qbit_metadata(
  main_item,
  files,
  pkgLock = NULL,
  qbit_env = NULL,
  packagesLoaded = names(sessionInfo()$otherPkgs),
  advertiseQBit = NULL,
  tstamp = Sys.time()
)
}
\arguments{
\item{main_item}{list;}

\item{files}{character;}

\item{pkgLock}{character; Package lock information from renv.lock}

\item{qbit_env}{environment;}

\item{packagesLoaded}{character; packages currently loaded}

\item{advertiseQBit}{logical; Define if qbit shall be advertised.}

\item{tstamp}{POSIXct; Time stamp to be used for QBit}
}
\value{
list containing metadata
}
\description{
Create metadata for Qbits as list objects.
}
