% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qbit_metadata.R
\name{create_qbit_metadata_path}
\alias{create_qbit_metadata_path}
\title{Create QBit Metadata}
\usage{
create_qbit_metadata_path(
  qbit_path,
  pattern_main = "main\\\\.R",
  index_file = file.path(qbit_path, "index.yml"),
  pkgLock = file.path(qbit_path, "renv.lock"),
  files_ignore = c("renv"),
  stage = Sys.getenv("STAGE", "dev")
)
}
\arguments{
\item{qbit_path}{character; Source path of qbit to be deployed.}

\item{pattern_main}{character; Path to qbit main file to be used.}

\item{index_file}{character; Path to meta data index file to be used.}

\item{pkgLock}{character; Path to package lock file.}

\item{files_ignore}{character; Files to be ignored.}

\item{stage}{character; Stage to be used for deployment}
}
\description{
Create QBit Metadata
}
