% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImagePath.R
\name{getImagePath}
\alias{getImagePath}
\title{Get Image path of QBit}
\usage{
getImagePath(img, stage = getOption("STAGE"))
}
\arguments{
\item{img}{character; File name of image.}

\item{stage}{character; Stage, either 'dev', or 'prod'}
}
\description{
Get Image path of QBit
}
