% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_image.R
\name{progress_image}
\alias{progress_image}
\title{Create Progress image for chapter}
\usage{
progress_image(
  title,
  idx_doing,
  idx_total,
  file_out,
  line_max_char = 24,
  file_badge = "badge.svg"
)
}
\description{
Create Progress image for chapter
}
\examples{
\dontrun{
progress_image("R is everywhere", 1, 6, "out.png")
progress_image("Use existing functions and data through packages", 5, 6, "out.png")
progress_image("Extract or replace columns in a data frame using `$`", 5, 6, "out.png")
}
}
