% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course_image.R
\name{course_image}
\alias{course_image}
\title{Create Progress image for chapter}
\usage{
course_image(
  title,
  file_out,
  line_max_char = 24,
  file_badge = "badge.svg",
  file_course = "main/course.svg"
)
}
\description{
Create Progress image for chapter
}
